+++ sdk/src/client/api/mana.rs
@@ -1,7 +1,13 @@
 // Copyright 2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::{client::Client, types::block::slot::SlotIndex};
+use crate::{
+    client::Client,
+    types::block::{
+        protocol::{lower_bits, multiplication_and_shift, upper_bits},
+        slot::SlotIndex,
+    },
+};
 
 impl Client {
     /// Calculates the potential mana that is generated by holding `amount` tokens from `slot_index_created` to
@@ -42,6 +48,18 @@ impl Client {
             let mana_generated = mana_structure.generate_mana(amount, slots_since_epoch_start as u32);
             mana_decayed + mana_generated
         } else {
+            let c = {
+                let amount_hi = upper_bits(amount);
+                let amount_lo = lower_bits(amount);
+                let (amount_hi, amount_lo) = multiplication_and_shift(
+                    amount_hi,
+                    amount_lo,
+                    mana_structure.decay_factor_epochs_sum() * mana_structure.generation_rate() as u32,
+                    mana_structure.decay_factor_epochs_sum_exponent() + mana_structure.generation_rate_exponent()
+                        - params.slots_per_epoch_exponent(),
+                );
+                amount_hi << 32 | amount_lo
+            };
             let slots_before_next_epoch = *slot_index_created
                 - **(epoch_index_created + 1)
                     .slot_index_range(slots_per_epoch_exponent.iter().copied())?
@@ -50,7 +68,6 @@ impl Client {
                 - **(epoch_index_target - 1)
                     .slot_index_range(slots_per_epoch_exponent.iter().copied())?
                     .end();
-            let c = params.generate_something(amount);
             let potential_mana_n = mana_structure.decay(
                 mana_structure.generate_mana(amount, slots_before_next_epoch as u32),
                 *epoch_index_target - *epoch_index_created,
